requirements.txt
asgiref==3.5.0
Django==4.0.3
django-cors-headers==3.11.0
django-dotenv==1.4.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
mysqlclient==2.1.0

